+++ PF/SistemaGestionBussiness/UsuarioBussiness.cs
@@ -18,8 +18,23 @@ namespace SistemaGestionBussiness
         public static Usuario GetUser(int userId)
         {
             UsuarioData usuarioData = new UsuarioData();
+            if (!usuarioData.UserExists(userId))
+            {
+                throw new ArgumentException($"No se encontró usuario con Id: {userId}");
+            }
 
             return usuarioData.GetUser(userId);
+        }
+
+        public static string GetUserNameById(int id)
+        {
+            UsuarioData usuarioData = new UsuarioData();
+            if (!usuarioData.UserExists(id))
+            {
+                throw new ArgumentException($"No se encontró usuario con Id: {id}");
+            }
+
+            return usuarioData.GetUserNameById(id);
         }
 
         public static bool CreateUser(Usuario user)
@@ -32,6 +47,10 @@ namespace SistemaGestionBussiness
         public static bool UpdateUser(int userId, Usuario user)
         {
             UsuarioData usuarioData = new UsuarioData();
+            if (!usuarioData.UserExists(userId))
+            {
+                throw new ArgumentException($"No se encontró usuario con Id: {userId}");
+            }
 
             return usuarioData.UpdateUser(userId, user);
         }
@@ -39,6 +58,10 @@ namespace SistemaGestionBussiness
         public static bool DeleteUser(int userId)
         {
             UsuarioData usuarioData = new UsuarioData();
+            if (!usuarioData.UserExists(userId))
+            {
+                throw new ArgumentException($"No se encontró usuario con Id: {userId}");
+            }
 
             return usuarioData.DeleteUser(userId);
         }
